Enable debuggin IW in vscode.web and add tests
