+++ .vscode/launch.json
@@ -201,12 +201,12 @@
                 "VSC_JUPYTER_FORCE_LOGGING": "1",
                 "VSC_JUPYTER_CI_TEST_GREP": "", // Leave as `VSCode Notebook` to run only Notebook tests.
                 "VSC_JUPYTER_CI_TEST_INVERT_GREP": "", // Initialize this to invert the grep (exclude tests with value defined in grep).
-                "CI_PYTHON_PATH": "", // Update with path to real python interpereter used for testing.
+                "CI_PYTHON_PATH": "/Users/donjayamanne/crap/.venv/bin/python", // Update with path to real python interpereter used for testing.
                 "VSC_JUPYTER_CI_RUN_NON_PYTHON_NB_TEST": "", // Initialize this to run tests again Julia & other kernels.
                 "VSC_JUPYTER_WEBVIEW_TEST_MIDDLEWARE": "true", // Initialize to create the webview test middleware
                 "VSC_JUPYTER_LOAD_EXPERIMENTS_FROM_FILE": "true",
                 // "TF_BUILD": "", // Set to anything to force full logging
-                "TEST_FILES_SUFFIX": "vscode.test",
+                "TEST_FILES_SUFFIX": "(*.vscode.test|*.vscode.common.test)",
                 "VSC_JUPYTER_REMOTE_NATIVE_TEST": "false", // Change to `true` to run the Native Notebook tests with remote jupyter connections.
                 "VSC_JUPYTER_NON_RAW_NATIVE_TEST": "false", // Change to `true` to run the Native Notebook tests with non-raw kernels (i.e. local jupyter server).
                 "XVSC_JUPYTER_INSTRUMENT_CODE_FOR_COVERAGE": "1",
