+++ pvsc.code-workspace
@@ -162,7 +162,7 @@
                     "VSC_JUPYTER_CI_RUN_NON_PYTHON_NB_TEST": "", // Initialize this to run tests again Julia & other kernels.
                     "VSC_JUPYTER_WEBVIEW_TEST_MIDDLEWARE": "true", // Initialize to create the webview test middleware
                     "VSC_JUPYTER_LOAD_EXPERIMENTS_FROM_FILE": "true",
-                    "TEST_FILES_SUFFIX": "vscode.test*",
+                    "TEST_FILES_SUFFIX": "(*.vscode.test|*.vscode.common.test)",
                     "VSC_JUPYTER_REMOTE_NATIVE_TEST": "false", // Change to `true` to run the Native Notebook tests with remote jupyter connections.
                     "VSC_JUPYTER_NON_RAW_NATIVE_TEST": "false", // Change to `true` to run the Native Notebook tests with non-raw kernels (i.e. local jupyter server).
                     "XVSC_JUPYTER_INSTRUMENT_CODE_FOR_COVERAGE": "1",
