+++ src/interactive-window/editor-integration/codelensprovider.ts
@@ -22,8 +22,8 @@ import { traceInfoIfCI } from '../../platform/logging';
 import {
     CodeLensCommands,
     EditorContexts,
-    PYTHON_FILE,
-    PYTHON_UNTITLED,
+    InteractiveInputScheme,
+    NotebookCellScheme,
     Telemetry
 } from '../../platform/common/constants';
 import { IDataScienceCodeLensProvider, ICodeWatcher } from './types';
@@ -78,7 +78,7 @@ export class DataScienceCodeLensProvider implements IDataScienceCodeLensProvider
     // CodeLensProvider interface
     // Some implementation based on DonJayamanne's jupyter extension work
     public provideCodeLenses(document: vscode.TextDocument, _token: vscode.CancellationToken): vscode.CodeLens[] {
-        if (document.uri.scheme != PYTHON_FILE.scheme && document.uri.scheme !== PYTHON_UNTITLED.scheme) {
+        if ([NotebookCellScheme, InteractiveInputScheme].includes(document.uri.scheme)) {
             return [];
         }
         // Get the list of code lens for this document.
