+++ src/test/datascience/interactiveDebugging.vscode.web.test.ts
@@ -1,0 +1,7 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+import { sharedIWDebuggerTests } from './interactiveDebugging.vscode.common';
+
+/* eslint-disable @typescript-eslint/no-explicit-any, no-invalid-this */
+suite('Interactive Window Debugging', sharedIWDebuggerTests);
