+++ src/test/datascience/notebook/remoteNotebookEditor.vscode.web.test.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-import { startJupyterServer } from './helper.web';
+import { startJupyterServer } from '../../common';
 import { sharedRemoteNotebookEditorTests } from './remoteNotebookEditor.vscode.common';
 
 suite('DataScience - VSCode Notebook - (Remote) (Execution) (slow)', function () {
