+++ src/test/web/index.web.test.ts
@@ -1,0 +1,6 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+import { initializeCommonWebApi } from '../common.web';
+
+initializeCommonWebApi();
